requirements.txt
streamlit
phidata
python-dotenv
anthropic
python-docx